adjust exploratory work outside the scope of fixing CVE-2025-3445
