+++ .goreleaser.yaml
@@ -28,7 +28,7 @@ builds:
       - -X github.com/derailed/k9s/cmd.version={{.Env.K9S_VERSION}}
       - -X github.com/google/go-containerregistry/cmd/crane/cmd.Version={{.Env.CRANE_VERSION}}
       - -X github.com/zarf-dev/zarf/src/cmd.syftVersion={{.Env.SYFT_VERSION}}
-      - -X github.com/zarf-dev/zarf/src/cmd.archivesVersion={{.Env.ARCHIVES_VERSION}}
+      - -X github.com/zarf-dev/zarf/src/cmd.archiverVersion={{.Env.ARCHIVER_VERSION}}
       - -X github.com/zarf-dev/zarf/src/cmd.helmVersion={{.Env.HELM_VERSION}}
     goarch:
       - amd64
