+++ README.md
@@ -99,7 +99,6 @@ To dive deeper into the tech, you can read the [Nerd Notes](https://docs.zarf.de
 We would also like to thank the following awesome libraries and projects without which Zarf would not be possible!
 
 [![pterm/pterm](https://img.shields.io/badge/pterm%2Fpterm-007d9c?logo=go&logoColor=white)](https://github.com/pterm/pterm)
-# FIXME(mkcp): make Archives
 [![mholt/archiver](https://img.shields.io/badge/mholt%2Farchiver-007d9c?logo=go&logoColor=white)](https://github.com/mholt/archiver)
 [![spf13/cobra](https://img.shields.io/badge/spf13%2Fcobra-007d9c?logo=go&logoColor=white)](https://github.com/spf13/cobra)
 [![go-git/go-git](https://img.shields.io/badge/go--git%2Fgo--git-007d9c?logo=go&logoColor=white)](https://github.com/go-git/go-git)
