+++ src/pkg/packager/sources/oci.go
@@ -193,6 +193,6 @@ func (s *OCISource) Collect(ctx context.Context, dir string) (string, error) {
 
 	_ = os.Remove(dstTarball)
 
-	// FIXME(mkcp): Support with internal/archive
+	// TODO(mkcp): See https://github.com/zarf-dev/zarf/issues/3051
 	return dstTarball, archiver.Archive(allTheLayers, dstTarball)
 }
