+++ src/pkg/packager/sources/url.go
@@ -7,7 +7,6 @@ package sources
 import (
 	"context"
 	"fmt"
-	"github.com/zarf-dev/zarf/src/internal/archive"
 	"os"
 	"path/filepath"
 	"strings"
@@ -49,7 +48,7 @@ func (s *URLSource) Collect(ctx context.Context, dir string) (string, error) {
 		return "", err
 	}
 
-	return archive.RenameFromMetadata(dstTarball)
+	return RenameFromMetadata(dstTarball)
 }
 
 // LoadPackage loads a package from an http, https or sget URL.
