+++ src/pkg/packager/sources/utils.go
@@ -6,10 +6,17 @@ package sources
 
 import (
 	"fmt"
+	"io"
+	"os"
+	"path/filepath"
 	"strings"
 
+	"github.com/defenseunicorns/pkg/helpers/v2"
+	goyaml "github.com/goccy/go-yaml"
+	"github.com/mholt/archiver/v3"
 	"github.com/zarf-dev/zarf/src/api/v1alpha1"
 	"github.com/zarf-dev/zarf/src/config"
+	"github.com/zarf-dev/zarf/src/pkg/layout"
 	"github.com/zarf-dev/zarf/src/pkg/zoci"
 )
 
@@ -27,6 +34,94 @@ func IsValidFileExtension(filename string) bool {
 	}
 
 	return false
+}
+
+func identifyUnknownTarball(path string) (string, error) {
+	if helpers.InvalidPath(path) {
+		return "", &os.PathError{Op: "open", Path: path, Err: os.ErrNotExist}
+	}
+	if filepath.Ext(path) != "" && IsValidFileExtension(path) {
+		return path, nil
+	} else if filepath.Ext(path) != "" && !IsValidFileExtension(path) {
+		return "", fmt.Errorf("%s is not a supported tarball format (%+v)", path, GetValidPackageExtensions())
+	}
+
+	// rename to .tar.zst and check if it's a valid tar.zst
+	tzst := fmt.Sprintf("%s.tar.zst", path)
+	if err := os.Rename(path, tzst); err != nil {
+		return "", err
+	}
+	// TODO(mkcp): See https://github.com/zarf-dev/zarf/issues/3051
+	format, err := archiver.ByExtension(tzst)
+	if err != nil {
+		return "", err
+	}
+	_, ok := format.(*archiver.TarZstd)
+	if ok {
+		return tzst, nil
+	}
+
+	// rename to .tar and check if it's a valid tar
+	tb := fmt.Sprintf("%s.tar", path)
+	if err := os.Rename(tzst, tb); err != nil {
+		return "", err
+	}
+	format, err = archiver.ByExtension(tb)
+	if err != nil {
+		return "", err
+	}
+	_, ok = format.(*archiver.Tar)
+	if ok {
+		return tb, nil
+	}
+
+	return "", fmt.Errorf("%s is not a supported tarball format (%+v)", path, GetValidPackageExtensions())
+}
+
+// RenameFromMetadata renames a tarball based on its metadata.
+func RenameFromMetadata(path string) (string, error) {
+	var pkg v1alpha1.ZarfPackage
+
+	ext := filepath.Ext(path)
+	if ext == "" {
+		pathWithExt, err := identifyUnknownTarball(path)
+		if err != nil {
+			return "", err
+		}
+		path = pathWithExt
+		ext = filepath.Ext(path)
+	}
+	if ext == ".zst" {
+		ext = ".tar.zst"
+	}
+
+	// TODO(mkcp): See https://github.com/zarf-dev/zarf/issues/3051
+	if err := archiver.Walk(path, func(f archiver.File) error {
+		if f.Name() == layout.ZarfYAML {
+			b, err := io.ReadAll(f)
+			if err != nil {
+				return err
+			}
+			if err := goyaml.Unmarshal(b, &pkg); err != nil {
+				return err
+			}
+		}
+		return nil
+	}); err != nil {
+		return "", err
+	}
+
+	if pkg.Metadata.Name == "" {
+		return "", fmt.Errorf("%q does not contain a zarf.yaml", path)
+	}
+
+	name := NameFromMetadata(&pkg, false)
+
+	name = fmt.Sprintf("%s%s", name, ext)
+
+	tb := filepath.Join(filepath.Dir(path), name)
+
+	return tb, os.Rename(path, tb)
 }
 
 // NameFromMetadata generates a name from a package's metadata.
