+++ src/test/e2e/08_create_differential_test.go
@@ -43,7 +43,7 @@ func TestCreateDifferential(t *testing.T) {
 	defer e2e.CleanFiles(t, differentialPackageName)
 
 	// Extract the yaml of the differential package
-	// FIXME(mkcp): Migrate to use internal/archive
+	// TODO(mkcp): See https://github.com/zarf-dev/zarf/issues/3051
 	err = archiver.Extract(differentialPackageName, layout.ZarfYAML, tmpdir)
 	require.NoError(t, err, "unable to extract zarf.yaml from the differential git package")
